+++ resources/views/frontend/shared-user-collections/show.blade.php
@@ -5,6 +5,14 @@
     @parent
 @stop
 
+@section('og')
+<meta property="og:title" content="{{ $collection->name }}" />
+<meta property="og:description" content="{{ $collection->description ?? trans('user-collection.shared.og.fallback_description') }}" />
+<meta property="og:type" content="website" />
+<meta property="og:url" content="{{ route('frontend.shared-user-collections.show', $collection) }}" />
+<meta property="og:image" content="{{ route('dielo.nahlad', ['id' => $items->first()->id, 'width' => 800]) }}" />
+@stop
+
 @section('main-navigation')
 <!-- Skip -->
 @stop
@@ -13,10 +21,10 @@
 <div class="container pt-5 mt-5">
     <div class="mb-5 pb-5 text-lg text-muted">
         <p>
-            Výber diel z <a href="#TODO" class="underline">Webu <i class="fa fa-search color"></i> umenia</a>.
+            Výber diel z <a href="{{ config('app.url') }}" class="underline">Webu <i class="fa fa-search color"></i> umenia</a>.
         </p>
         <p class="mt-3">
-            Inšpirované výstavou <a href="#TODO" class="underline">Sedmičky v SNG</a>.
+            Inšpirované výstavou <a href="https://medium.com/sng-online/vystava-sedmicky/home" class="underline">Sedmičky v SNG</a>.
         </p>
     </div>
     <x-user-collections.share-form
@@ -26,10 +34,10 @@
     />
     <div class="pb-5 my-5 text-center text-lg text-muted">
         <p>
-            Vytvor a zdieľaj vlastný výber výtvarných diel na <a href="#TODO" class="underline">Webe umenia</a>.
+            Vytvor a zdieľaj vlastný výber výtvarných diel na <a href="{{ config('app.url') }}" class="underline">Webe umenia</a>.
         </p>
         <p class="mt-3">
-            Tematické výbery kurátoriek a kurátorov Slovenskej národnej galérie nájdeš na výstave <a href="#TODO" class="underline">Sedmičky</a>.
+            Tematické výbery kurátoriek a kurátorov Slovenskej národnej galérie nájdeš na výstave <a href="https://medium.com/sng-online/vystava-sedmicky/home" class="underline">Sedmičky</a>.
         </p>
     </div>
 </div>