+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+
+## [2.32.1] - 2021-10-14
+### Fixed
+- public views for shared user collections
+- pagination for articles and edu articles
+
 ## [2.32.0] - 2021-10-04
 ### Added
 - paging for front page, articles and edu articles